整理scheduler_func，独立元气值相关页面和函数
