+++ app/views.py
@@ -1628,373 +1628,6 @@ def modpw(request):
     return render(request, "modpw.html", locals())
 
 
-# 用已有的搜索，加一个转账的想他转账的 field
-# 调用的时候传一下 url 到 origin
-# 搜索不希望出现学号，rid 为 User 的 index
-@login_required(redirect_field_name="origin")
-@utils.check_user_access(redirect_url="/logout/")
-@log.except_captured(source='views[transaction_page]', record_user=True)
-def transaction_page(request, rid=None):
-    valid, user_type, html_display = utils.check_user_type(request.user)
-    me = utils.get_person_or_org(request.user, user_type)
-    html_display["is_myself"] = True
-    html_display['warn_code'] = 0
-
-    try:
-        user = User.objects.get(id=rid)
-        recipient = utils.get_person_or_org(user)
-
-    except:
-        return redirect(
-            "/welcome/?warn_code=1&warn_message=该用户不存在，无法实现转账!")
-    if not hasattr(recipient, "organization_id") or user_type != "Organization":
-        html_display = wrong("目前只支持小组向小组转账！")
-    if request.user == user:
-        html_display=wrong("不能向自己转账！")
-    if html_display['warn_code']==1:
-        if hasattr(recipient, "organization_id"):
-            return redirect(
-                "/orginfo/?name={name}&warn_code=1&warn_message={message}".format(
-                    name=recipient.oname,
-                    message=html_display['warn_message']))
-        else:
-            return  redirect(
-                "/stuinfo/?name={name}&warn_code=1&warn_message={message}".format(
-                    name=recipient.name,
-                    message=html_display['warn_message']))
-    # 新版侧边栏, 顶栏等的呈现，采用 bar_display, 必须放在render前最后一步
-    # 如果希望前移，请联系YHT
-    bar_display = utils.get_sidebar_and_navbar(request.user)
-    bar_display["title_name"] = "Transaction"
-    bar_display["navbar_name"] = "发起转账"
-
-    # 获取名字
-    _, _, context = utils.check_user_type(user)
-    context = utils.get_sidebar_and_navbar(user, bar_display=context)
-    name = recipient.oname
-    context["name"] = name
-    context["rid"] = rid
-    context["YQPoint"] = me.YQPoint
-
-    # 如果是post, 说明发起了一起转账
-    # 到这里, rid没有问题, 接收方和发起方都已经确定
-    if request.method == "POST":
-        # 获取转账消息, 如果没有消息, 则为空
-        transaction_msg = request.POST.get("msg", "")
-
-        # 检查发起转账的数据
-        try:
-            amount = float(request.POST["amount"])
-            assert amount > 0
-            assert int(amount * 10) == amount * 10
-        except:
-            return redirect(message_url(wrong('非法的转账数量!')))
-
-        # 到这里, 参数的合法性检查完成了, 接下来应该是检查发起人的账户, 够钱就转
-        try:
-            notification = None
-            with transaction.atomic():
-                # 首先锁定用户
-                payer = (
-                    Organization.objects.activated()
-                        .select_for_update()
-                        .get(organization_id=request.user)
-                )
-
-                # 接下来确定金额
-                if payer.oname != YQP_ONAME and payer.YQPoint < amount:
-                    html_display["warn_code"] = 1
-                    html_display["warn_message"] = (
-                            "现存元气值余额为"
-                            + str(payer.YQPoint)
-                            + ", 不足以发起额度为"
-                            + str(amount)
-                            + "的转账!"
-                    )
-                else:
-                    payer.YQPoint -= amount
-                    record = TransferRecord.objects.create(
-                        proposer=request.user,
-                        recipient=user,
-                        amount=amount,
-                        message=transaction_msg,
-                        rtype=TransferRecord.TransferType.TRANSACTION
-                    )
-                    record.save()
-                    payer.save()
-                    warn_message = "成功发起向" + name + "的转账! 元气值将在对方确认后到账。"
-
-                    content_msg = transaction_msg if transaction_msg else f'转账金额：{amount}'
-                    notification = notification_create(
-                        receiver=user,
-                        sender=request.user,
-                        typename=Notification.Type.NEEDDO,
-                        title=Notification.Title.TRANSFER_CONFIRM,
-                        content=content_msg,
-                        URL="/myYQPoint/",
-                        relate_TransferRecord=record,
-                    )
-            if notification is not None:
-                publish_notification(
-                    notification,
-                    app=WechatApp.TRANSFER,
-                    level=WechatMessageLevel.IMPORTANT,
-                )
-            return redirect("/myYQPoint/")
-
-        except Exception as e:
-            # print(e)
-            html_display["warn_code"] = 1
-            html_display["warn_message"] = "出现无法预料的问题, 请联系管理员!"
-
-
-    # 新版侧边栏, 顶栏等的呈现，采用 bar_display, 必须放在render前最后一步
-    # 如果希望前移，请联系YHT
-    bar_display = utils.get_sidebar_and_navbar(request.user)
-    bar_display["title_name"] = "Transaction"
-    bar_display["navbar_name"] = "发起转账"
-    return render(request, "transaction_page.html", locals())
-
-
-
-@log.except_captured(source='views[confirm_transaction]', record_user=True)
-def confirm_transaction(request, tid=None, reject=None):
-    context = dict()
-    context["warn_code"] = 1  # 先假设有问题
-    new_notification = None
-    with transaction.atomic():
-        try:
-            record = TransferRecord.objects.select_for_update().get(
-                id=tid, recipient=request.user
-            )
-
-        except Exception as e:
-
-            context["warn_message"] = "交易遇到问题, 请联系管理员!" + str(e)
-            return context
-
-        if record.status != TransferRecord.TransferStatus.WAITING:
-            context["warn_message"] = "交易已经完成, 请不要重复操作!"
-            return context
-
-        payer = record.proposer
-        try:
-            if hasattr(payer, "naturalperson"):
-                payer = (
-                    NaturalPerson.objects.activated()
-                        .select_for_update()
-                        .get(person_id=payer)
-                )
-            else:
-                payer = Organization.objects.select_for_update().get(
-                    organization_id=payer
-                )
-        except:
-            context["warn_message"] = "交易对象不存在或已毕业, 请联系管理员!"
-            return context
-
-        recipient = record.recipient
-        if hasattr(recipient, "naturalperson"):
-            recipient = (
-                NaturalPerson.objects.activated()
-                    .select_for_update()
-                    .get(person_id=recipient)
-            )
-        else:
-            recipient = Organization.objects.select_for_update().get(
-                organization_id=recipient
-            )
-
-        if reject is True:
-            record.status = TransferRecord.TransferStatus.REFUSED
-            payer.YQPoint += record.amount
-            payer.save()
-            context["warn_message"] = "拒绝转账成功!"
-            new_notification = notification_create(
-                receiver=record.proposer,
-                sender=record.recipient,
-                typename=Notification.Type.NEEDREAD,
-                title=Notification.Title.TRANSFER_FEEDBACK,
-                content=f"{str(recipient)}拒绝了您的转账。",
-                URL="/myYQPoint/",
-            )
-            notification_status_change(record.transfer_notification.get().id)
-        else:
-            record.status = TransferRecord.TransferStatus.ACCEPTED
-            recipient.YQPoint += record.amount
-            recipient.save()
-            context["warn_message"] = "交易成功!"
-            new_notification = notification_create(
-                receiver=record.proposer,
-                sender=record.recipient,
-                typename=Notification.Type.NEEDREAD,
-                title=Notification.Title.TRANSFER_FEEDBACK,
-                content=f"{str(recipient)}接受了您的转账。",
-                URL="/myYQPoint/",
-            )
-            notification_status_change(record.transfer_notification.get().id)
-        publish_notification(new_notification, app=WechatApp.TRANSFER)
-        record.finish_time = datetime.now()  # 交易完成时间
-        record.save()
-        context["warn_code"] = 2
-
-        return context
-
-    context["warn_message"] = "交易遇到问题, 请联系管理员!"
-    return context
-
-
-@log.except_captured(source='views[record2Display]')
-def record2Display(record_list, user):  # 对应myYQPoint函数中的table_show_list
-    lis = []
-    amount = {"send": 0.0, "recv": 0.0}
-    # 储存这个列表中所有record的元气值的和
-    for record in record_list:
-        lis.append({})
-
-        # 确定类型
-        record_type = "send" if record.proposer.username == user.username else "recv"
-
-        # id
-        lis[-1]["id"] = record.id
-
-        # 时间
-        lis[-1]["start_time"] = record.start_time.strftime("%Y-%m-%d %H:%M")
-        if record.finish_time is not None:
-            lis[-1]["finish_time"] = record.finish_time.strftime("%Y-%m-%d %H:%M")
-
-        # 对象
-        # 如果是给出列表，那么对象就是接收者
-
-        obj_user = record.recipient if record_type == "send" else record.proposer
-        lis[-1]["obj_direct"] = "To  " if record_type == "send" else "From"
-        if hasattr(obj_user, "naturalperson"):  # 如果OneToOne Field在个人上
-            lis[-1]["obj"] = obj_user.naturalperson.name
-            lis[-1]["obj_url"] = "/stuinfo/?name=" + lis[-1]["obj"] + "+" + str(obj_user.id)
-        else:
-            lis[-1]["obj"] = obj_user.organization.oname
-            lis[-1]["obj_url"] = "/orginfo/?name=" + lis[-1]["obj"]
-
-        # 金额
-        lis[-1]["amount"] = record.amount
-        amount[record_type] += record.amount
-
-        # 留言
-        lis[-1]["message"] = record.message
-        lis[-1]["if_act_url"] = False
-        if record.corres_act is not None:
-            lis[-1]["message"] = "报名活动" + record.corres_act.title
-            # TODO 这里还需要补充一个活动跳转链接
-
-        # 状态
-        if record.status == TransferRecord.TransferStatus.PENDING:
-            # PENDING 就不对个人可见了，个人看到的就是元气值已经转过去了
-            lis[-1]["status"] = "已接收"
-        else:
-            lis[-1]["status"] = record.get_status_display()
-
-    # 对外展示为 1/10
-    """
-    统一在前端修改
-    for key in amount:
-        amount[key] = amount[key]/10
-    """
-    # 由于误差, 将amount调整为小数位数不超过2
-    for key in amount.keys():
-        amount[key] = round(amount[key], 1)
-    return lis, amount
-
-
-# modified by Kinnuch
-
-
-@login_required(redirect_field_name="origin")
-@utils.check_user_access(redirect_url="/logout/")
-@log.except_captured(source='views[myYQPoint]', record_user=True)
-def myYQPoint(request):
-    valid, user_type, html_display = utils.check_user_type(request.user)
-
-    # 接下来处理POST相关的内容
-    html_display["warn_code"] = 0
-    if request.method == "POST":  # 发生了交易处理的事件
-        try:  # 检查参数合法性
-            post_args = request.POST.get("post_button")
-            record_id, action = post_args.split("+")[0], post_args.split("+")[1]
-            assert action in ["accept", "reject"]
-            reject = action == "reject"
-        except:
-            html_display["warn_code"] = 1
-            html_display["warn_message"] = "交易遇到问题,请不要非法修改参数!"
-
-        if html_display["warn_code"] == 0:  # 如果传入参数没有问题
-            # 调用确认预约API
-            context = confirm_transaction(request, record_id, reject)
-            # 此时warn_code一定是1或者2，必定需要提示
-            html_display["warn_code"] = context["warn_code"]
-            html_display["warn_message"] = context["warn_message"]
-
-    me = utils.get_person_or_org(request.user, user_type)
-    html_display["is_myself"] = True
-
-    to_send_set = TransferRecord.objects.filter(
-        proposer=request.user, status=TransferRecord.TransferStatus.WAITING
-    )
-
-    to_recv_set = TransferRecord.objects.filter(
-        recipient=request.user, status=TransferRecord.TransferStatus.WAITING
-    )
-
-    issued_send_set = TransferRecord.objects.filter(
-        proposer=request.user,
-        status__in=[
-            TransferRecord.TransferStatus.ACCEPTED,
-            TransferRecord.TransferStatus.REFUSED,
-            # PENDING 目前只用于个人报名预报备活动时使用
-            TransferRecord.TransferStatus.PENDING,
-        ],
-    )
-
-    issued_recv_set = TransferRecord.objects.filter(
-        recipient=request.user,
-        status__in=[
-            TransferRecord.TransferStatus.ACCEPTED,
-            TransferRecord.TransferStatus.REFUSED,
-        ],
-    )
-
-    # to_set 按照开始时间降序排列
-    to_set = to_send_set.union(to_recv_set).order_by("-start_time")
-    # issued_set 按照完成时间及降序排列
-    # 这里应当要求所有已经issued的记录是有执行时间的
-    issued_set = issued_send_set.union(issued_recv_set).order_by("-finish_time")
-
-    to_list, amount = record2Display(to_set, request.user)
-    issued_list, _ = record2Display(issued_set, request.user)
-    send_list = []
-    receive_list = []
-    for item in issued_list:
-        if item["obj_direct"] == "To  ":
-            send_list.append(item)
-        else:
-            receive_list.append(item)
-
-    show_table = {
-        "obj": "对象",
-        "time": "时间",
-        "amount": "金额",
-        "message": "留言",
-        "status": "状态",
-    }
-
-    # 新版侧边栏, 顶栏等的呈现，采用 bar_display, 必须放在render前最后一步
-    bar_display = utils.get_sidebar_and_navbar(request.user, "我的元气值")
-    # 补充一些呈现信息
-    # bar_display["title_name"] = "My YQPoint"
-    # bar_display["navbar_name"] = "我的元气值"  #
-    # bar_display["help_message"] = local_dict["help_message"]["我的元气值"]
-
-    return render(request, "myYQPoint.html", locals())
-
 
 @login_required(redirect_field_name="origin")
 @utils.check_user_access(redirect_url="/logout/")
@@ -2052,6 +1685,7 @@ def saveShowPositionStatus(request):
             position.show_post = False
         position.save()
     return JsonResponse({"success": True})
+
 
 @login_required(redirect_field_name="origin")
 @utils.check_user_access(redirect_url="/logout/")
